gemini-adk-agent/requirements.txt
flask==3.0.3
google-generativeai==0.7.2
gunicorn==22.0.0
